requirements.txt
attrs==21.4.0
CacheMan==2.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cyhunspell==2.0.2
filelock==3.6.0
future==0.18.2
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
nltk==3.7
numpy==1.22.3
packaging==21.3
psutil==5.9.0
pyparsing==3.0.8
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
six==1.16.0
syntok==1.4.4
tokenizers==0.11.6
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.9
